+++ CHANGES.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 - add break update portfolio when hive market is down
 - separate DEC rental payments in cost and earning
   - NOTE best to remove all season_* files and re-capture all seasonal information
+- Add Modern Wild filter to main page
 
 ## 0.4.0
 - Remove previous migrations